fix: compile and clippy fixes for rustc 1.98
rustc 1.98 tightened const-generic checking on type aliases and clippy
grew the manual_div_ceil / unnecessary_cast lints, breaking the stable
CI leg:

- PolyRingVector/PolyRingMatrix aliases declared DEGREE_BOUND as u64
  while PolyRing requires usize; align them to usize
- replace the four hand-rolled (x+7)/8 div_ceil8 bodies with
  usize::div_ceil (stable since 1.73, fine under the 1.85 MSRV)
- drop redundant `bits as usize` casts in mldsa encoding

Verified locally on both CI toolchains (1.98.1 and 1.85.0):
fmt/clippy/build/test/bench all green, 479 tests passing.

+++ crates/algebra/src/matrix/matrix.rs
@@ -25,7 +25,7 @@ pub type RingMatrix<R> = GenericMatrix<R>;
 pub type PolynomialMatrix<R> = GenericMatrix<R>;
 
 /// A matrix over a polynomial ring R[x]/(x^d+1)
-pub type PolyRingMatrix<R, const DEGREE_BOUND: u64> = GenericMatrix<PolyRing<R, DEGREE_BOUND>>;
+pub type PolyRingMatrix<R, const DEGREE_BOUND: usize> = GenericMatrix<PolyRing<R, DEGREE_BOUND>>;
 
 /// Threshold for parallel operations (row count)
 #[cfg(feature = "parallel")]
